requirements.txt
streamlit==1.18.1
azure-identity
azure-keyvault-secrets
snowflake-snowpark-python==1.0.0
reportlab==3.6.12
pandas==1.4.4
pyarrow==8.0.0
PyPDF2==3.0.1
streamlit-javascript==0.1.5
streamlit-aggrid==0.3.3
streamlit-toggle-switch==1.0.2
streamlit-toggle==0.1.3